+++ app/assets/config/wilday_ui_manifest.js
@@ -1,1 +1,4 @@
 //= link_directory ../stylesheets/wilday_ui .css
+//= link_directory ../stylesheets/wilday_ui/components .css
+//= link_directory ../stylesheets/wilday_ui/components/button .css
+//= link_directory ../stylesheets/wilday_ui/components/button/features .css