+++ lib/wilday_ui.rb
@@ -1,5 +1,6 @@
 require "wilday_ui/version"
 require "wilday_ui/engine"
+require "wilday_ui/config/theme"
 
 module WildayUi
   # Your code goes here...