+++ test/dummy/app/controllers/components_controller.rb
@@ -1,4 +1,0 @@
-class ComponentsController < ApplicationController
-  def button
-  end
-end